feat: add ability to list scheduled jobs
In an effort to solve the fact that the topic bot currently posts and
changes topics multiple times per day, I've added a way to check the
existing schedule() jobs as well as increased the default poller timer.

My theory is that the poller was being called before schedule could
declare a job done. My hope is the 60S timer will be enough to solve
this. If not, more research will need to happen